requirements.txt
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
coinbase-advanced-py>=1.8.0
flask>=3.0.0
